requirement.txt
numpy>=1.20.0
numba>=0.52.0
scipy>=1.2.3
mujoco==2.3.0
casadi==3.5.5
matplotlib==3.3.3
opencv-python==4.5.1.48
pybullet==3.0.8
qpsolvers
tqdm==4.56.0
quadprog
pin
pyyaml
robosuite==1.4.0
robomimic==0.3.0
gym
wandb
h5py
torch==2.1.2
torchvision==0.16.2
torchaudio
pynput
pyrealsense==2.53.1.4623
hidapi
